+++ lib/tests/nagware.test.ts
@@ -2,20 +2,15 @@
  * @jest-environment node
  */
 
-import { NagLevel, NagwareSubmission } from "@lib/types";
+import { NagLevel, NagwareSubmission, VaultStatus } from "@lib/types";
 import { detectOldUnprocessedSubmissions } from "@lib/nagware";
 import { getAppSetting } from "@lib/appSettings";
 
 jest.mock("@lib/appSettings");
 
-enum MockStatus {
-  New,
-  Downloaded,
-}
-
-function mockOldSubmission(status: MockStatus, numberOfDaysInThePast: number): NagwareSubmission {
+function mockOldSubmission(status: VaultStatus, numberOfDaysInThePast: number): NagwareSubmission {
   return {
-    status: status === MockStatus.New ? "New" : "Downloaded",
+    status: status,
     createdAt: Date.now() - numberOfDaysInThePast * (1000 * 60 * 60 * 24),
   };
 }
@@ -45,10 +40,10 @@ describe("Nagware - detectOldUnprocessedSubmissions", () => {
 
     it("There are multiple submissions under 22 days old", async () => {
       const submissions = [
-        mockOldSubmission(MockStatus.New, 3),
-        mockOldSubmission(MockStatus.Downloaded, 10),
-        mockOldSubmission(MockStatus.New, 21),
-        mockOldSubmission(MockStatus.Downloaded, 21),
+        mockOldSubmission(VaultStatus.NEW, 3),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 10),
+        mockOldSubmission(VaultStatus.NEW, 21),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 21),
       ];
 
       const sut = await detectOldUnprocessedSubmissions(submissions);
@@ -61,9 +56,9 @@ describe("Nagware - detectOldUnprocessedSubmissions", () => {
   describe("Should return NagLevel.UnsavedSubmissionsOver21DaysOld when", () => {
     it("There is 1 new submission between 22 and 35 days old", async () => {
       const submissions = [
-        mockOldSubmission(MockStatus.New, 3),
-        mockOldSubmission(MockStatus.New, 10),
-        mockOldSubmission(MockStatus.New, 22),
+        mockOldSubmission(VaultStatus.NEW, 3),
+        mockOldSubmission(VaultStatus.NEW, 10),
+        mockOldSubmission(VaultStatus.NEW, 22),
       ];
 
       const sut = await detectOldUnprocessedSubmissions(submissions);
@@ -74,9 +69,9 @@ describe("Nagware - detectOldUnprocessedSubmissions", () => {
 
     it("There are multiple new submissions between 22 and 35 days old", async () => {
       const submissions = [
-        mockOldSubmission(MockStatus.New, 3),
-        mockOldSubmission(MockStatus.New, 25),
-        mockOldSubmission(MockStatus.New, 22),
+        mockOldSubmission(VaultStatus.NEW, 3),
+        mockOldSubmission(VaultStatus.NEW, 25),
+        mockOldSubmission(VaultStatus.NEW, 22),
       ];
 
       const sut = await detectOldUnprocessedSubmissions(submissions);
@@ -87,10 +82,10 @@ describe("Nagware - detectOldUnprocessedSubmissions", () => {
 
     it("There are multiple submissions (new and downloaded) between 22 and 35 days old", async () => {
       const submissions = [
-        mockOldSubmission(MockStatus.New, 3),
-        mockOldSubmission(MockStatus.New, 25),
-        mockOldSubmission(MockStatus.Downloaded, 25),
-        mockOldSubmission(MockStatus.New, 22),
+        mockOldSubmission(VaultStatus.NEW, 3),
+        mockOldSubmission(VaultStatus.NEW, 25),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 25),
+        mockOldSubmission(VaultStatus.NEW, 22),
       ];
 
       const sut = await detectOldUnprocessedSubmissions(submissions);
@@ -103,9 +98,9 @@ describe("Nagware - detectOldUnprocessedSubmissions", () => {
   describe("Should return NagLevel.UnconfirmedSubmissionsOver21DaysOld when", () => {
     it("There is 1 downloaded submission between 22 and 35 days old", async () => {
       const submissions = [
-        mockOldSubmission(MockStatus.Downloaded, 3),
-        mockOldSubmission(MockStatus.Downloaded, 10),
-        mockOldSubmission(MockStatus.Downloaded, 22),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 3),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 10),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 22),
       ];
 
       const sut = await detectOldUnprocessedSubmissions(submissions);
@@ -116,9 +111,9 @@ describe("Nagware - detectOldUnprocessedSubmissions", () => {
 
     it("There are multiple downloaded submissions between 22 and 35 days old", async () => {
       const submissions = [
-        mockOldSubmission(MockStatus.Downloaded, 3),
-        mockOldSubmission(MockStatus.Downloaded, 25),
-        mockOldSubmission(MockStatus.Downloaded, 22),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 3),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 25),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 22),
       ];
 
       const sut = await detectOldUnprocessedSubmissions(submissions);
@@ -129,10 +124,10 @@ describe("Nagware - detectOldUnprocessedSubmissions", () => {
 
     it("There are multiple new submissions (under 21 days old) and downloaded submissions (between 22 and 35 days old)", async () => {
       const submissions = [
-        mockOldSubmission(MockStatus.New, 3),
-        mockOldSubmission(MockStatus.New, 21),
-        mockOldSubmission(MockStatus.Downloaded, 25),
-        mockOldSubmission(MockStatus.New, 21),
+        mockOldSubmission(VaultStatus.NEW, 3),
+        mockOldSubmission(VaultStatus.NEW, 21),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 25),
+        mockOldSubmission(VaultStatus.NEW, 21),
       ];
 
       const sut = await detectOldUnprocessedSubmissions(submissions);
@@ -145,9 +140,9 @@ describe("Nagware - detectOldUnprocessedSubmissions", () => {
   describe("Should return NagLevel.UnsavedSubmissionsOver35DaysOld when", () => {
     it("There is 1 new submission over 35 days old", async () => {
       const submissions = [
-        mockOldSubmission(MockStatus.New, 3),
-        mockOldSubmission(MockStatus.New, 10),
-        mockOldSubmission(MockStatus.New, 36),
+        mockOldSubmission(VaultStatus.NEW, 3),
+        mockOldSubmission(VaultStatus.NEW, 10),
+        mockOldSubmission(VaultStatus.NEW, 36),
       ];
 
       const sut = await detectOldUnprocessedSubmissions(submissions);
@@ -158,10 +153,10 @@ describe("Nagware - detectOldUnprocessedSubmissions", () => {
 
     it("There are multiple new submissions over 35 days old", async () => {
       const submissions = [
-        mockOldSubmission(MockStatus.New, 3),
-        mockOldSubmission(MockStatus.New, 25),
-        mockOldSubmission(MockStatus.New, 36),
-        mockOldSubmission(MockStatus.New, 40),
+        mockOldSubmission(VaultStatus.NEW, 3),
+        mockOldSubmission(VaultStatus.NEW, 25),
+        mockOldSubmission(VaultStatus.NEW, 36),
+        mockOldSubmission(VaultStatus.NEW, 40),
       ];
 
       const sut = await detectOldUnprocessedSubmissions(submissions);
@@ -172,11 +167,11 @@ describe("Nagware - detectOldUnprocessedSubmissions", () => {
 
     it("There are multiple new or downloaded submissions (between 1 and 35 days old) and 1 new submission over 35 days old", async () => {
       const submissions = [
-        mockOldSubmission(MockStatus.New, 3),
-        mockOldSubmission(MockStatus.Downloaded, 25),
-        mockOldSubmission(MockStatus.Downloaded, 35),
-        mockOldSubmission(MockStatus.New, 22),
-        mockOldSubmission(MockStatus.New, 36),
+        mockOldSubmission(VaultStatus.NEW, 3),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 25),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 35),
+        mockOldSubmission(VaultStatus.NEW, 22),
+        mockOldSubmission(VaultStatus.NEW, 36),
       ];
 
       const sut = await detectOldUnprocessedSubmissions(submissions);
@@ -189,9 +184,9 @@ describe("Nagware - detectOldUnprocessedSubmissions", () => {
   describe("Should return NagLevel.UnconfirmedSubmissionsOver35DaysOld when", () => {
     it("There is 1 downloaded submission over 35 days old", async () => {
       const submissions = [
-        mockOldSubmission(MockStatus.Downloaded, 3),
-        mockOldSubmission(MockStatus.Downloaded, 10),
-        mockOldSubmission(MockStatus.Downloaded, 36),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 3),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 10),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 36),
       ];
 
       const sut = await detectOldUnprocessedSubmissions(submissions);
@@ -202,10 +197,10 @@ describe("Nagware - detectOldUnprocessedSubmissions", () => {
 
     it("There are multiple downloaded submissions over 35 days old", async () => {
       const submissions = [
-        mockOldSubmission(MockStatus.Downloaded, 3),
-        mockOldSubmission(MockStatus.Downloaded, 25),
-        mockOldSubmission(MockStatus.Downloaded, 36),
-        mockOldSubmission(MockStatus.Downloaded, 40),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 3),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 25),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 36),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 40),
       ];
 
       const sut = await detectOldUnprocessedSubmissions(submissions);
@@ -216,11 +211,11 @@ describe("Nagware - detectOldUnprocessedSubmissions", () => {
 
     it("There are multiple new or downloaded submissions (between 1 and 35 days old) and 1 downloaded submission over 35 days old", async () => {
       const submissions = [
-        mockOldSubmission(MockStatus.New, 3),
-        mockOldSubmission(MockStatus.Downloaded, 25),
-        mockOldSubmission(MockStatus.Downloaded, 35),
-        mockOldSubmission(MockStatus.New, 22),
-        mockOldSubmission(MockStatus.Downloaded, 36),
+        mockOldSubmission(VaultStatus.NEW, 3),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 25),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 35),
+        mockOldSubmission(VaultStatus.NEW, 22),
+        mockOldSubmission(VaultStatus.DOWNLOADED, 36),
       ];
 
       const sut = await detectOldUnprocessedSubmissions(submissions);